Adding elements to array ohne ende

+++ tracker_project/tracker/models.py
@@ -1,5 +1,5 @@
-from django.db import models
 
+from django.db import models
 
 class BMI(models.Model):
     age = models.PositiveIntegerField(blank = True, null = False)
@@ -10,7 +10,7 @@ class BMI(models.Model):
     weightKilos = models.FloatField(blank = True, null=True)
     
 class Weight(models.Model):
-    mass = models.FloatField(blank = True, null = True)
+    mass = models.PositiveIntegerField(blank = True, null = True)
     day = models.PositiveIntegerField(blank = True, null=True)
     month = models.PositiveIntegerField(blank = True, null=True)
     year = models.PositiveIntegerField(blank = True, null=False)
